requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
click==6.4                # via pip-tools
defusedxml==0.4.1
django-debug-toolbar==1.4
django-filter==0.13.0
django-tagging==0.4.1
django-tastypie==0.13.3
django==1.9.4
djangorestframework==3.3.3
first==2.0.1              # via pip-tools
lxml==3.6.0
pip-tools==1.6
psycopg2==2.6.1
python-dateutil==2.5.2    # via django-tastypie
python-mimeparse==1.5.1   # via django-tastypie
pyyaml==3.11
six==1.10.0               # via pip-tools, python-dateutil
sqlparse==0.1.19          # via django-debug-toolbar
https://github.com/unbit/uwsgi/archive/uwsgi-2.0.zip#egg=uwsgi
